+++ src/main/java/com/elibrary/services/CategoryService.java
@@ -1,16 +1,20 @@
 package com.elibrary.services;
 
+import com.elibrary.Exception.BusinessNotFound;
+import com.elibrary.Exception.CategoryException;
 import com.elibrary.dto.request.CategoryRequest;
 import com.elibrary.dto.response.CategoryResponse;
 import com.elibrary.model.entity.Category;
 import com.elibrary.model.repos.CategoryRepo;
 import com.elibrary.model.specification.CategorySpecification;
+import org.checkerframework.checker.nullness.Opt;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
@@ -47,20 +51,29 @@ public class CategoryService {
         return category.map(this::convertCategoryToResponse).orElse(null);
     }
     
-    public CategoryResponse addCategory(CategoryRequest request){
+    public CategoryResponse addCategory(CategoryRequest request) throws CategoryException {
+        if(categoryRepo.existsByCategory(request.getCategory())){
+            throw new CategoryException("Category already exists");
+        }
         Category category = new Category();
         category.setCategory(request.getCategory());
         categoryRepo.save(category);
         return convertCategoryToResponse(category);
     }
 
-    public CategoryResponse updateCategory(Long id,CategoryRequest request){
-        CategoryResponse categoryResponse = findById(id);
-        Category category =  new Category();
-        category.setId(categoryResponse.getId());
-        category.setCategory(request.getCategory());
-        categoryRepo.save(category);
-        return convertCategoryToResponse(category);
+    public CategoryResponse updateCategory(Long id,CategoryRequest request) throws BusinessNotFound, CategoryException {
+        Optional<Category> category = categoryRepo.findById(id);
+        if(category.isPresent()){
+            Category category1 = category.get();
+            if(categoryRepo.existsByCategory(request.getCategory()) && !Objects.equals(categoryRepo.findByCategory(request.getCategory()).getId(), id)){
+                throw new CategoryException("Category already exists");
+            }
+            category1.setCategory(request.getCategory());
+            categoryRepo.save(category1);
+            return convertCategoryToResponse(category1);
+        }else {
+            throw new BusinessNotFound("Category not found");
+        }
     }
 
     public List<CategoryResponse> findAll(String search){
@@ -71,8 +84,14 @@ public class CategoryService {
         )).collect(Collectors.toList());
     }
 
-    public void delete(long id){
-        categoryRepo.deleteById(id);
+    public void delete(long id) throws BusinessNotFound {
+        Optional<Category> category = categoryRepo.findById(id);
+        if (category.isPresent()){
+            categoryRepo.deleteById(id);
+        }else {
+            throw new BusinessNotFound("Category not found");
+        }
+
     }
 
 }
